Added two extension methods as a Animation Curve replacement
ToPointArray
Converts an animation curve into an array of floats. Takes samples at even spots along the graph, combines the extra samples with the original data and sorts them by their graph time into an array.

BlobArray.Evaluate
Gets a value from a converted animation curve by taking a 0-1 value, converting it to a decimal index of the array, separates the integer and decimal of the decimal index, returns the result of lerping between the 2 closest integer indexes by the decimal.