Created an entity spawning system queue
Created a spawning system that allows any system to get a singleton queue and request that an entity prefab be spawned at the beginning of the next frame.
It uses two components; SpawnerData holds static baked information for the spawner system to reference (prefabs, configs etc.), SpawnerQueue holds the NativeQueue that is used for requesting spawns.

To add a new entity prefab you must create multiple things;
In SpawnerData;
	- add an entity variable
	- add an Enum entry
	- add a line in GetEntityPrefab connecting the Enum and entity

In Extensions add an entry into ToString for the Enum

In the spawner system you can also add a preset for the struct (like PlayerPreset)