Finished Attribute System (i hope)
The attribute system allows the use of requests and queues to Add, Edit or Remove Attributes and Attribute Mods. Attributes are values that can be used by other systems to influence gameplay, like health, damage, acceleration speed etc. Attributes have a base value and can have Attribute Mods. Mods can apply addition, multiplication and exponentiation to the base value in order to create scale in gameplay. Mods can also have a limited lifetime. When adding a mod a ModRefPointer is provided which can be use to manually alter the mod later. Attributes calculate their modified value and place it in the ModValue counterpart component, that component is what should be used by other systems. There are also AttributeRequests that can be sent in order to alter a ModValue.Value, this would most likely only be used for damaging entities.